Refactors core object structure
Moves core classes into a single `objects.py` file.
