+++ src/easydiffraction/experiments/experiment.py
@@ -13,7 +13,7 @@ from easydiffraction.experiments.iterable_components.datastore import DatastoreF
 
 from easydiffraction.utils.formatting import paragraph, warning
 from easydiffraction.utils.chart_plotter import ChartPlotter
-from easydiffraction.core.datablock import Datablock
+from easydiffraction.core.objects import Datablock
 from easydiffraction.core.constants import (DEFAULT_SAMPLE_FORM,
                                             DEFAULT_BEAM_MODE,
                                             DEFAULT_RADIATION_PROBE,
