+++ src/easydiffraction/experiments/iterable_components/background.py
@@ -7,10 +7,10 @@ from scipy.interpolate import interp1d
 
 from easydiffraction.utils.formatting import (paragraph,
                                               warning)
-from easydiffraction.core.parameter import (Parameter,
-                                            Descriptor)
-from easydiffraction.core.component import StandardComponent
-from easydiffraction.core.collection import Collection
+from easydiffraction.core.objects import (Parameter,
+                                            Descriptor,
+                                          StandardComponent,
+                                          Collection)
 from easydiffraction.core.constants import DEFAULT_BACKGROUND_TYPE
