+++ src/easydiffraction/experiments/standard_components/experiment_type.py
@@ -1,5 +1,5 @@
-from easydiffraction.core.component import StandardComponent
-from easydiffraction.core.parameter import Descriptor
+from easydiffraction.core.objects import (Descriptor,
+StandardComponent)
 
 
 class ExperimentType(StandardComponent):
