+++ src/easydiffraction/sample_models/sample_models.py
@@ -3,10 +3,10 @@ from abc import ABC, abstractmethod
 from easydiffraction.sample_models.standard_components.space_group import SpaceGroup
 from easydiffraction.sample_models.standard_components.cell import Cell
 from easydiffraction.sample_models.iterable_components.atom_sites import AtomSites
-from easydiffraction.core.collection import Collection
+from easydiffraction.core.objects import (Collection,
+Datablock)
 from easydiffraction.utils.formatting import paragraph
 from easydiffraction.crystallography import crystallography as ecr
-from easydiffraction.core.datablock import Datablock
 
 
 class SampleModel(Datablock):
